stock prediction
An LSTM-based deep learning model for predicting stock prices. This repository contains the code and data for training and evaluating a Long Short-Term Memory (LSTM) neural network to make stock price predictions. The model is designed to learn patterns and trends in historical stock price data and generate forecasts for future price movements.